requirements.txt
fastapi==0.104.1
json5==0.9.14
keras==2.14.0
nltk==3.8.1
numpy==1.25.2
pandas==2.1.2
psycopg2==2.9.9
python-dotenv==1.0.0
scikit-learn==1.3.1
spacy==3.7.2
tensorflow==2.14.0
uvicorn==0.23.2